+++ Diary/Data/Extensions/Future+.swift
@@ -1,0 +1,23 @@
+//
+//  Future+.swift
+//  Diary
+//
+//  Created by Jaehun Lee on 5/16/24.
+//
+
+import Combine
+
+extension Future where Failure == Error {
+    convenience init(operation: @escaping () async throws -> Output) {
+        self.init { promise in
+            Task {
+                do {
+                    let output = try await operation()
+                    promise(.success(output))
+                } catch {
+                    promise(.failure(error))
+                }
+            }
+        }
+    }
+}
