+++ Diary/Data/PersistentStorage/CoreDataStorage/CoreDataStorage.swift
@@ -7,12 +7,6 @@
 
 import CoreData
 
-enum CoreDataStorageError: Error {
-    case readError(Error)
-    case saveError(Error)
-    case deleteError(Error)
-}
-
 final class CoreDataStorage {
     static let shared = CoreDataStorage()
