+++ Diary/Data/Repositories/DiaryRepository.swift
@@ -43,18 +43,3 @@ final class DiaryRepository: DiaryRepositoryProtocol {
         .eraseToAnyPublisher()
     }
 }
-
-extension Future where Failure == Error {
-    convenience init(operation: @escaping () async throws -> Output) {
-        self.init { promise in
-            Task {
-                do {
-                    let output = try await operation()
-                    promise(.success(output))
-                } catch {
-                    promise(.failure(error))
-                }
-            }
-        }
-    }
-}
