--- Diary/Helper/Constants.swift
@@ -1,12 +1,0 @@
-//
-//  Constants.swift
-//  Diary
-//
-//  Created by Prism, Gama on 4/30/24.
-//
-
-import Foundation
-
-enum Constants {
-    static let jsonFileName = "sample"
-}
