+++ Diary/Presentations/Extensions/FormatStyle+.swift
