--- Diary/Resources/Assets.xcassets/sample.dataset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "data" : [
-    {
-      "filename" : "sample.json",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
